Remove function service call
