+++ core-ui/src/components/Lambdas/gql/hooks/queries/index.js
@@ -1,4 +1,3 @@
 export * from './useConfigMapQuery';
 export * from './useLambdaQuery';
 export * from './useLambdasQuery';
-export * from './useServiceQuery';
