+++ core-ui/src/components/Lambdas/gql/hooks/queries/testMocks/index.js
@@ -4,4 +4,3 @@ export * from './useLambdaQuery';
 export * from './useLambdasQuery';
 export * from './useServiceBindingUsagesQuery';
 export * from './useServiceInstancesQuery';
-export * from './useServiceQuery';
